requirements.txt
coverage==3.7.1
Django==2.1.5
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==11.0.0
django-bootstrap4==0.0.7
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
pytz==2018.9
selenium==2.53.1
sqlparse==0.2.1
